FIX: compiler warnings (VC++)
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@3103 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/libs/Base/cmtkActiveShapeModel.cxx
@@ -207,7 +207,7 @@ ActiveShapeModel::Decompose
     w[mode] = (deviation * *thisMode) / thisMode->EuclidNorm();
     
     const Types::Coordinate variance = (*(this->ModeVariances))[mode];
-    pdf *= exp( -(w[mode]*w[mode]) / (2.0 * variance) ) / sqrt( 2.0 * M_PI * variance);
+    pdf *= static_cast<float>( exp( -(w[mode]*w[mode]) / (2.0 * variance) ) / sqrt( 2.0 * M_PI * variance) );
     }
 #else
   float distance = 0.0;
